SecondMetaDataCanPullSurvivalAsMission Green
Note: This is not a good step of TDD - I am rewriting already working code to be cleaner, and using TDD in the process. I am skipping a few steps in the name of effeciency by stealing some low level code (for what I already know works and is clean) to speed up the process.